Move enums to lib, configure rollup export
