+++ .eslintrc
@@ -17,7 +17,7 @@
     "jsx-a11y/control-has-associated-label": "off",
     "import/extensions": "off",
     "import/no-extraneous-dependencies": "off",
-    "import/no-unresolved": "off",
+    "import/no-unresolved": "error",
     "import/prefer-default-export": "off",
     "no-restricted-syntax": "off",
     "react/no-typos": "off",
@@ -30,7 +30,21 @@
     "react/state-in-constructor": "off",
     "react/jsx-filename-extension": "off",
     "react/prop-types": "off",
-    "quotes": [2, "single", { "avoidEscape": true }]
+    "quotes": [2, "single", { "avoidEscape": true }],
+    "no-use-before-define": "off",
+    "@typescript-eslint/no-use-before-define": ["error"],
+    "@typescript-eslint/ban-ts-comment": "off",
+    "@typescript-eslint/no-empty-function": "off"
+  },
+  "settings": {
+    "import/parsers": {
+      "@typescript-eslint/parser": [".ts", ".tsx"]
+    },
+    "import/resolver": {
+      "typescript": {
+        "project": "./tsconfig.json"
+      }
+    }
   },
   "globals": {
     "fetch": true,
