+++ src/lib/index.js
@@ -1,2 +1,2 @@
-export { useForm } from "src/lib/useForm.ts";
-export { Form } from "src/lib/Form";
+export { useForm } from './useForm';
+export { Form } from './Form';
